requirements.txt
streamlit
python-dotenv
PyPDF2
pandas
langchain
langchain-google-genai
